hackcanada-model/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
Flask>=3.0.0
pandas>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
Werkzeug>=3.0.0
python-dotenv>=1.0.0
